requirements.txt
# FastAPI Framework
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1

# Web Scraping
playwright==1.58.0
playwright-stealth==1.0.6
firecrawl-py>=1.0.0
setuptools==75.8.0

# AI Agent
pydantic-ai==1.63.0
anthropic==0.84.0
openai==2.24.0

# Notifications
python-telegram-bot>=22.6

# Configuration
pyyaml==6.0.3
python-dotenv>=1.2.2

# Async & Utilities
httpx>=0.28.1
aiofiles==25.1.0

# Logging
structlog==25.5.0
logfire==4.25.0

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1

# Development Tools
black==26.1.0
ruff==0.15.4
mypy==1.19.1
